#6: rename ImageBackground to BackgroundImage
